accounting: Modify behavior of fs.CountError
Fixes #5865

fs.CountError is called when an error is encountered.
The method was calling GlobalStats().Error(err) which incremented the
error at the global stats level.
This led to calls to core/stats with group= filter returning an error
count of 0 even if errors actually occured.

This change requires the context to be provided when calling
fs.CountError. Doing so, we can retrieve the correct StatsInfo to
increment the errors from.