check and cryptcheck: report directory differences with --report-dirs - fixes #6440
Before this change, check and cryptcheck would not inform the user of directory
differences (for example, an empty directory that exists only on the src.)

After this change, a new --report-dirs flag allows alerting users of such
differences.

Note that directories will be reported regardless of whether they are empty.
They will be included in the total error count, but counted and summarized
separately, and not included in output files. --report-dirs is ignored when
--checkfile is in use.